requirements.txt
Flask-PyMongo
flask-redis
requests
responses
python-slugify